+++ BackEnd_GameArena/src/main/java/com/unifucamp/gamearena/controller/InscricaoController.java
@@ -1,6 +1,6 @@
 package com.unifucamp.gamearena.controller;
 
-import com.unifucamp.gamearena.controller.dto.InscricaoDTO;
+import com.unifucamp.gamearena.dto.InscricaoDTO;
 import com.unifucamp.gamearena.domain.Inscricao;
 import com.unifucamp.gamearena.mapper.InscricaoMapper;
 import com.unifucamp.gamearena.service.InscricaoService;
@@ -8,6 +8,7 @@ import jakarta.validation.Valid;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -23,12 +24,13 @@ public class InscricaoController {
         this.inscricaoService = inscricaoService;
     }
 
-    @PostMapping
+    @PostMapping("/registrar")
     public ResponseEntity<InscricaoDTO> criarInscricao(@RequestBody @Valid InscricaoDTO inscricaoDTO) {
         Inscricao inscricaoSalva = inscricaoService.salvar(InscricaoMapper.dtoToDomain(inscricaoDTO));
         return ResponseEntity.ok(InscricaoMapper.domainToDTO(inscricaoSalva));
     }
 
+    @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
     @PutMapping("/{id}")
     public ResponseEntity<InscricaoDTO> atualizarInscricao(@PathVariable Integer id, @RequestBody InscricaoDTO inscricaoDTO) {
         return inscricaoService.atualizar(id, InscricaoMapper.dtoToDomain(inscricaoDTO))
@@ -36,6 +38,7 @@ public class InscricaoController {
                 .orElse(ResponseEntity.notFound().build());
     }
 
+    @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
     @GetMapping("/{id}")
     public ResponseEntity<InscricaoDTO> buscarPorId(@PathVariable Integer id) {
         return inscricaoService.buscarPorId(id)
@@ -56,6 +59,7 @@ public class InscricaoController {
         return ResponseEntity.ok(inscricoesDTO);
     }
 
+    @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> desativarInscricao(@PathVariable Integer id) {
         boolean desativado = inscricaoService.desativar(id);
