+++ BackEnd_GameArena/src/main/java/com/unifucamp/gamearena/domain/Inscricao.java
