+++ BackEnd_GameArena/src/main/java/com/unifucamp/gamearena/repository/InscricaoRepository.java
@@ -1,6 +1,6 @@
 package com.unifucamp.gamearena.repository;
 
-import com.unifucamp.gamearena.repository.entity.InscricaoEntity;
+import com.unifucamp.gamearena.entity.InscricaoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InscricaoRepository extends JpaRepository<InscricaoEntity, Integer> {
